Refactoring: move simulationinputs to top level of simulation
