+++ src/app/features/simulator/services/simulation-dialogue.service.ts
@@ -5,7 +5,7 @@ import {SimulatorComponent} from "@app/features/simulator/components/simulator/s
 import {MatDialog} from "@angular/material/dialog";
 import {InfoTrustDetailsComponent} from "@app/features/simulator/components/info-trust-details/info-trust-details.component";
 import {SimulationInputDetails} from "@app/features/simulator/components/simulation-input-details/simulation-input-details.component";
-import {SimulationInputs} from "@app/features/simulator/models/simulation-inputs";
+import {SimulationInputs} from "@app/features/simulator/simulation-inputs";
 import {TrustComputationService} from "@app/features/simulator/services/trust-computation.service";
 
 @Injectable({
