+++ src/app/features/simulator/services/trust-computation.service.spec.ts
@@ -6,7 +6,7 @@ import {Author} from "@app/shared/keml/core/author";
 import {InformationLinkType} from "@app/shared/keml/json/knowledge-models";
 import {ConversationJson} from "@app/shared/keml/json/sequence-diagram-models";
 import {JsonFixer} from "@app/shared/keml/json2core/json-fixer";
-import {SimulationInputs} from "@app/features/simulator/models/simulation-inputs";
+import {SimulationInputs} from "@app/features/simulator/simulation-inputs";
 import {TestBed} from "@angular/core/testing";
 
 describe('TrustComputationService', () => {
