+++ src/app/features/simulator/simulation-inputs.ts
